Unified Easy and Medium quizzes to be generated based on difficulty. Updated routing in App.js
